+++ nucliadb/src/nucliadb/ingest/orm/processor/processor.py
@@ -52,16 +52,10 @@ from nucliadb.ingest.orm.processor.data_augmentation import (
 )
 from nucliadb.ingest.orm.resource import (
     Resource,
-    compute_resource_status,
-    delete_basic_computedmetadata_classifications,
-    extract_field_metadata_languages,
-    get_text_field_mimetype,
-    maybe_update_basic_icon,
-    maybe_update_basic_summary,
-    maybe_update_basic_thumbnail,
-    update_basic_computedmetadata_classifications,
-    update_basic_languages,
 )
+from nucliadb_models import content_types
+from nucliadb_models.common import CloudLink
+from nucliadb_models.content_types import GENERIC_MIME_TYPE
 from nucliadb_protos import (
     knowledgebox_pb2,
     resources_pb2,
@@ -821,57 +815,55 @@ def _merge_basic(
     Merge an incoming basic payload onto the existing basic, applying all
     deduplication and immutability rules, then return the result ready to persist.
     """
-    if not current.ByteSize():
-        # Brand-new resource — use payload as-is (after field deletions).
-        merged = PBBasic()
-        merged.CopyFrom(payload)
-    else:
-        merged = PBBasic()
-        merged.CopyFrom(current)
+    assert current.ByteSize(), "Expected an existing resource with non-empty basic"
+    merged = PBBasic()
+    merged.CopyFrom(current)
 
-        if merged != payload:
-            for field in _BASIC_IMMUTABLE_FIELDS:
-                # Immutable basic fields that are already set are cleared from
-                # the payload so that they are not overwritten.
-                if getattr(merged, field, "") != "":
-                    payload.ClearField(field)  # type: ignore[arg-type]
+    if merged != payload:
+        for field in _BASIC_IMMUTABLE_FIELDS:
+            # Immutable basic fields that are already set are cleared from
+            # the payload so that they are not overwritten.
+            if getattr(merged, field, "") != "":
+                payload.ClearField(field)  # type: ignore[arg-type]
 
-            merged.MergeFrom(payload)
+        merged.MergeFrom(payload)
 
-            # Prevent duplicated languages
-            unique_languages = set(merged.metadata.languages)
-            merged.metadata.ClearField("languages")
-            merged.metadata.languages.extend(unique_languages)
+        # Prevent duplicated languages
+        unique_languages = set(merged.metadata.languages)
+        merged.metadata.ClearField("languages")
+        merged.metadata.languages.extend(unique_languages)
 
-            # Prevent duplicated labels
-            unique_labels = set(merged.labels)
-            merged.ClearField("labels")
-            merged.labels.extend(unique_labels)
+        # Prevent duplicated labels
+        unique_labels = set(merged.labels)
+        merged.ClearField("labels")
+        merged.labels.extend(unique_labels)
 
-            # Update processing status: only overwrite when the payload signals
-            # that the resource is useful (i.e. coming from a real processing run).
-            if payload.HasField("metadata") and payload.metadata.useful:
-                merged.metadata.status = payload.metadata.status
+        # Update processing status: only overwrite when the payload signals
+        # that the resource is useful (i.e. coming from a real processing run).
+        if payload.HasField("metadata") and payload.metadata.useful:
+            merged.metadata.status = payload.metadata.status
 
-            # We force the usermetadata classifications to be exactly what the
-            # payload defines (no merging).
-            if payload.HasField("usermetadata"):
-                merged.usermetadata.CopyFrom(payload.usermetadata)
+        # We force the usermetadata classifications to be exactly what the
+        # payload defines (no merging).
+        if payload.HasField("usermetadata"):
+            merged.usermetadata.CopyFrom(payload.usermetadata)
 
-            if len(payload.fieldmetadata):
-                # Keep only the last fieldmetadata item for each field. API
-                # users are responsible to manage fieldmetadata properly.
-                fields: list[str] = []
-                positions: dict[str, int] = {}
-                for i, fieldmetadata in enumerate(merged.fieldmetadata):
-                    field_id = f"{FIELD_TYPE_PB_TO_STR[fieldmetadata.field.field_type]}/{fieldmetadata.field.field}"
-                    if field_id not in fields:
-                        fields.append(field_id)
-                    positions[field_id] = i
+        if len(payload.fieldmetadata):
+            # Keep only the last fieldmetadata item for each field. API
+            # users are responsible to manage fieldmetadata properly.
+            fields: list[str] = []
+            positions: dict[str, int] = {}
+            for i, fieldmetadata in enumerate(merged.fieldmetadata):
+                field_id = (
+                    f"{FIELD_TYPE_PB_TO_STR[fieldmetadata.field.field_type]}/{fieldmetadata.field.field}"
+                )
+                if field_id not in fields:
+                    fields.append(field_id)
+                positions[field_id] = i
 
-                updated = [merged.fieldmetadata[positions[f]] for f in fields]
-                del merged.fieldmetadata[:]
-                merged.fieldmetadata.extend(updated)
+            updated = [merged.fieldmetadata[positions[f]] for f in fields]
+            del merged.fieldmetadata[:]
+            merged.fieldmetadata.extend(updated)
 
     # Some basic fields are computed off field metadata.
     # Recompute upon field deletions.
@@ -940,3 +932,191 @@ def _apply_extracted_basic_updates(
         extracted_languages.extend(extract_field_metadata_languages(fcmw))
 
     update_basic_languages(basic, extracted_languages)
+
+
+async def compute_resource_status(
+    txn: Transaction,
+    kbid: str,
+    uuid: str,
+    basic: PBBasic,
+) -> None:
+    """
+    Compute and set the resource-level processing status on basic by inspecting
+    the status of all individual fields.
+    """
+    field_ids = await datamanagers.resources.get_all_field_ids(
+        txn, kbid=kbid, rid=uuid, for_update=False
+    )
+    if field_ids is None:
+        # No fields, it is processed
+        basic.metadata.status = resources_pb2.Metadata.Status.PROCESSED
+        return
+
+    field_statuses = await datamanagers.fields.get_statuses(
+        txn, kbid=kbid, rid=uuid, fields=field_ids.fields
+    )
+
+    # If any field is processing -> PENDING
+    if any(f.status == writer_pb2.FieldStatus.Status.PENDING for f in field_statuses):
+        basic.metadata.status = resources_pb2.Metadata.Status.PENDING
+    # If we have any non-DA error -> ERROR
+    elif any(
+        f.status == writer_pb2.FieldStatus.Status.ERROR
+        and any(
+            e.source_error.severity == writer_pb2.Error.Severity.ERROR
+            and e.source_error.code != writer_pb2.Error.ErrorCode.DATAAUGMENTATION
+            for e in f.errors
+        )
+        for f in field_statuses
+    ):
+        basic.metadata.status = resources_pb2.Metadata.Status.ERROR
+    # Otherwise (everything processed or we only have DA errors) -> PROCESSED
+    else:
+        basic.metadata.status = resources_pb2.Metadata.Status.PROCESSED
+
+
+def delete_basic_computedmetadata_classifications(basic: PBBasic, deleted_fields: list[FieldID]) -> bool:
+    """
+    We keep a copy of field classifications computed by the processing engine at the basic object
+    so that users can easily access them without having to load the field metadata from the storage.
+
+    This funcion removes the field classifications for the fields that have been deleted.
+    Returns whether the basic was modified.
+    """
+    if len(deleted_fields) == 0:
+        # Nothing to delete
+        return False
+    new_field_classifications = [
+        fc for fc in basic.computedmetadata.field_classifications if fc.field not in deleted_fields
+    ]
+    if len(new_field_classifications) == len(basic.computedmetadata.field_classifications):
+        # No changes
+        return False
+
+    basic.computedmetadata.ClearField("field_classifications")
+    basic.computedmetadata.field_classifications.extend(new_field_classifications)
+    return True
+
+
+def maybe_update_basic_summary(basic: PBBasic, summary_text: str) -> bool:
+    if basic.summary or not summary_text:
+        return False
+    basic.summary = summary_text
+    return True
+
+
+def maybe_update_basic_icon(basic: PBBasic, mimetype: str | None) -> bool:
+    if basic.icon not in (None, "", "application/octet-stream", GENERIC_MIME_TYPE):
+        # Icon already set or detected
+        return False
+
+    if not mimetype:
+        return False
+
+    if not content_types.valid(mimetype):
+        logger.warning(
+            "Invalid mimetype. Skipping icon update.",
+            extra={"mimetype": mimetype, "rid": basic.uuid, "slug": basic.slug},
+        )
+        return False
+
+    basic.icon = mimetype
+    return True
+
+
+def maybe_update_basic_thumbnail(
+    basic: PBBasic, thumbnail: resources_pb2.CloudFile | None, kbid: str
+) -> bool:
+    if basic.thumbnail or thumbnail is None:
+        return False
+    basic.thumbnail = CloudLink.format_reader_download_uri(thumbnail.uri)
+    fix_kbid_in_thumbnail(basic, kbid)
+    return True
+
+
+def extract_field_metadata_languages(
+    field_metadata: resources_pb2.FieldComputedMetadataWrapper,
+) -> list[str]:
+    languages: set[str] = set()
+    languages.add(field_metadata.metadata.metadata.language)
+    for _, splitted_metadata in field_metadata.metadata.split_metadata.items():
+        languages.add(splitted_metadata.language)
+    return list(languages)
+
+
+PB_TEXT_FORMAT_TO_MIMETYPE = {
+    resources_pb2.FieldText.Format.PLAIN: "text/plain",
+    resources_pb2.FieldText.Format.HTML: "text/html",
+    resources_pb2.FieldText.Format.RST: "text/x-rst",
+    resources_pb2.FieldText.Format.MARKDOWN: "text/markdown",
+    resources_pb2.FieldText.Format.JSON: "application/json",
+    resources_pb2.FieldText.Format.KEEP_MARKDOWN: "text/markdown",
+    resources_pb2.FieldText.Format.JSONL: "application/x-ndjson",
+    resources_pb2.FieldText.Format.PLAIN_BLANKLINE_SPLIT: "text/plain+blankline",
+}
+
+
+def get_text_field_mimetype(bm: writer_pb2.BrokerMessage) -> str | None:
+    if len(bm.texts) == 0:
+        return None
+    text_format = next(iter(bm.texts.values())).format
+    return PB_TEXT_FORMAT_TO_MIMETYPE[text_format]
+
+
+def update_basic_languages(basic: writer_pb2.Basic, languages: list[str]) -> bool:
+    if len(languages) == 0:
+        return False
+
+    updated = False
+    for language in languages:
+        if not language:
+            continue
+
+        if basic.metadata.language == "":
+            basic.metadata.language = language
+            updated = True
+
+        if language not in basic.metadata.languages:
+            basic.metadata.languages.append(language)
+            updated = True
+
+    return updated
+
+
+def update_basic_computedmetadata_classifications(
+    basic: PBBasic, fcmw: resources_pb2.FieldComputedMetadataWrapper
+) -> bool:
+    """
+    We keep a copy of field classifications computed by the processing engine at the basic object
+    so that users can easily access them without having to load the field metadata from the storage.
+
+    This function updates the basic object with the new field computed metadata.
+    Returns whether the basic was modified.
+    """
+    some_deleted = delete_basic_computedmetadata_classifications(basic, [fcmw.field])
+
+    fcfs = resources_pb2.FieldClassifications()
+    fcfs.field.CopyFrom(fcmw.field)
+
+    some_added = False
+    if len(fcmw.metadata.metadata.classifications) > 0:
+        some_added = True
+        fcfs.classifications.extend(fcmw.metadata.metadata.classifications)
+
+    for split_id, split in fcmw.metadata.split_metadata.items():
+        if split_id not in fcmw.metadata.deleted_splits:
+            if len(split.classifications) > 0:
+                some_added = True
+                fcfs.classifications.extend(split.classifications)
+    if some_added:
+        basic.computedmetadata.field_classifications.append(fcfs)
+    return some_added or some_deleted
+
+
+def fix_kbid_in_thumbnail(basic: PBBasic, kbid: str):
+    if basic.thumbnail.startswith("/kb/") and not basic.thumbnail.startswith(f"/kb/{kbid}/"):
+        # Replace the kbid in the thumbnail if it doesn't match the current kbid. This is necessary for
+        # resources that have been backed up and we are restoring them to a different kbid.
+        parts = basic.thumbnail.split("/", 3)
+        parts[2] = kbid
+        basic.thumbnail = "/".join(parts)
